requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.0
frozenlist==1.7.0
h11==0.16.0
idna==3.10
livekit-protocol==1.0.4
multidict==6.6.3
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.46.2
types-protobuf==6.30.2.20250703
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
yarl==1.20.1
